requirements-dev.txt
attrs
cloudpickle
coverage
databroker
doct
doctr
flake8
ipython
ipywidgets
jinja2
lmfit
matplotlib != 3.0.1  # bug in scatter method
mongoquery
networkx
nose
numpydoc
ophyd
pandas
pims
pyepics
pyqt5
pytest
pytest-faulthandler
pyyaml
pyzmq
requests
scipy
sphinx
sphinx_rtd_theme
streamz
# These suitcases are test deps of databroker which we need to access
# databroker fixtures.
suitcase-jsonl
suitcase-mongo
suitcase-msgpack
tifffile
